refactor: Add GenericPublisher
Instead of overriding publish task, implement proper PublisherInterface.
